+++ packages/dart/lib/src/objects/response/parse_error_response.dart
@@ -5,7 +5,8 @@ ParseResponse buildErrorResponse(
     ParseResponse response, ParseNetworkResponse apiResponse) {
   final Map<String, dynamic> responseData = json.decode(apiResponse.data);
   response.error = ParseError(
-      code: responseData[keyCode], message: responseData[keyError].toString());
-  response.statusCode = responseData[keyCode];
+      code: responseData[keyCode] ?? ParseError.otherCause,
+      message: responseData[keyError].toString());
+  response.statusCode = responseData[keyCode] ?? ParseError.otherCause;
   return response;
 }