+++ packages/dart/test/parse_encoder_test.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:parse_server_sdk/parse_server_sdk.dart';
 import 'package:test/test.dart';
 
@@ -8,7 +10,7 @@ void main() {
     // arrange
     await Parse().initialize(
       'appId',
-      'https://test.parse.com',
+      'https://example.com',
       debug: true,
       // to prevent automatic detection
       fileDirectory: 'someDirectory',
@@ -20,10 +22,10 @@ void main() {
       appVersion: 'someAppVersion',
     );
 
+    // act
     ParseObject parseObject2 = ParseObject("objectId2");
     parseObject2.objectId = "objectId2";
 
-    // List and Map
     parseObject2
         .setAdd("dataParseObjectList", ["ListText1", "ListText2", "ListText3"]);
     parseObject2.setAdd("dataParseObjectMap", {
@@ -32,19 +34,44 @@ void main() {
       'KeyTestMap3': 'ValueTestMap3',
     });
 
-    // parseObject2 inside parseObject1
     ParseObject parseObject1 = ParseObject("parseObject1");
     parseObject1.objectId = "objectId1";
     parseObject1.setAdd("dataParseObject2", parseObject2);
 
-    // desired output
-    String expectedResult =
-        "{className: parseObject1, objectId: objectId1, dataParseObject2: {__op: Add, objects: [{className: objectId2, objectId: objectId2, dataParseObjectList: {__op: Add, objects: [[ListText1, ListText2, ListText3]]}, dataParseObjectMap: {__op: Add, objects: [{KeyTestMap1: ValueTestMap1, KeyTestMap2: ValueTestMap2, KeyTestMap3: ValueTestMap3}]}}]}}";
-
-    // act
     dynamic actualResult = parseEncode(parseObject1, full: true);
 
-    //assert
-    expect(actualResult.toString(), expectedResult);
+    var objectDesiredOutput = {
+      "className": "parseObject1",
+      "objectId": "objectId1",
+      "dataParseObject2": {
+        "__op": "Add",
+        "objects": [
+          {
+            "className": "objectId2",
+            "objectId": "objectId2",
+            "dataParseObjectList": {
+              "__op": "Add",
+              "objects": [
+                ["ListText1", "ListText2", "ListText3"]
+              ],
+            },
+            "dataParseObjectMap": {
+              "__op": "Add",
+              "objects": [
+                {
+                  "KeyTestMap1": "ValueTestMap1",
+                  "KeyTestMap2": "ValueTestMap2",
+                  "KeyTestMap3": "ValueTestMap3"
+                }
+              ]
+            }
+          }
+        ],
+      },
+    };
+    var objectJsonDesiredOutput = jsonEncode(objectDesiredOutput);
+
+    // assert
+    expect(jsonEncode(actualResult), objectJsonDesiredOutput);
   });
 }