+++ packages/flutter/test/parse_client_configuration_test.dart
@@ -6,6 +6,7 @@ void main() {
   SharedPreferences.setMockInitialValues(<String, String>{});
 
   test('testBuilder', () async {
+    // arrange
     await Parse().initialize('appId', 'serverUrl',
         clientKey: 'clientKey',
         liveQueryUrl: 'liveQueryUrl',
@@ -17,6 +18,7 @@ void main() {
         fileDirectory: 'someDirectory',
         debug: true);
 
+    // assert
     expect(ParseCoreData().applicationId, 'appId');
     expect(ParseCoreData().serverUrl, 'serverUrl');
     expect(ParseCoreData().clientKey, 'clientKey');